Added javadoc and refactored in own package
