+++ smslib-camel/src/main/java/org/smslib/camel/SMSLibComponent.java
@@ -1,4 +1,4 @@
-package org.smslib;
+package org.smslib.camel;
 
 import java.util.Map;
 
@@ -9,6 +9,11 @@ import org.slf4j.LoggerFactory;
 
 /**
  * Represents the component that manages {@link SMSLibEndpoint}.
+ * Two "paths" are supported: <code>mock</code> and <code>modem</code>.
+ * <p><code>mock</code> uses the {@link org.smslib.gateway.MockGateway} via the {@link MockEndpoint} to provide a dummy interface to test with.</p>
+ * <p><code>modem</code> uses the SMSlib based serial modem API via the {@link ModemEndpoint} to use real GSM hardware with camel.
+* <p><b>Note:</b> The main focus of this endpoint is to received messages. Therefore the sending capabilities are limited. Sess {@link SMSLibProducer} for details.</p>
+ * @author derjust
  */
 public class SMSLibComponent extends UriEndpointComponent {
 	public SMSLibComponent() {
