+++ smslib-camel/src/main/java/org/smslib/camel/SMSLibProducer.java
@@ -1,4 +1,4 @@
-package org.smslib;
+package org.smslib.camel;
 
 import org.apache.camel.Exchange;
 import org.apache.camel.impl.DefaultProducer;
@@ -7,10 +7,12 @@ import org.slf4j.LoggerFactory;
 import org.smslib.gateway.AbstractGateway;
 import org.smslib.message.OutboundMessage;
 
-import static org.smslib.SMSLibEndpoint.*;
+import static org.smslib.camel.SMSLibEndpoint.*;
 
 /**
  * The SMSLib producer.
+ * It is quite simple and expects the body to be a String and the header id {@link SMSLibEndpoint#HEADER_RECIPIENT_NUMBER} containing a number 
+ * @author derjust
  */
 public class SMSLibProducer extends DefaultProducer {
     private static final Logger LOG = LoggerFactory.getLogger(SMSLibProducer.class);
