+++ smslib-camel/src/main/resources/META-INF/services/org/apache/camel/component/smslib
@@ -1,1 +1,1 @@
-class=org.smslib.SMSLibComponent
+class=org.smslib.camel.SMSLibComponent
