+++ smslib-camel/src/test/java/org/smslib/SMSLibMockTest.java
@@ -5,6 +5,13 @@ import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.camel.test.junit4.CamelTestSupport;
 import org.junit.Test;
 
+/**
+ * Test case to verify that the bare minimum setup (SMSlib's mock interface) 
+ * is working in Camel scenario.
+ * 
+ * @author derjust
+ *
+ */
 public class SMSLibMockTest extends CamelTestSupport {
 
     @Test
@@ -19,7 +26,7 @@ public class SMSLibMockTest extends CamelTestSupport {
     protected RouteBuilder createRouteBuilder() throws Exception {
         return new RouteBuilder() {
             public void configure() {
-                from("smslib://mock?delay=10000")
+                from("smslib://mock?delay=1000")
                   .to("mock:result");
             }
         };
